cursor-project/wechat-tool/requirements.txt
# 微信机器人相关
itchat-uos==1.5.0.dev0
wechatpy==1.8.15

# 文件处理相关
python-docx==0.8.11
PyPDF2==3.0.1
Pillow==10.0.1
reportlab==4.0.4
docx2txt==0.8

# 图片处理
opencv-python==4.8.1.78
img2pdf==0.5.1

# 定时任务
schedule==1.2.0
APScheduler==3.10.4

# 新闻API
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 配置管理
python-dotenv==1.0.0
pyyaml==6.0.1

# 日志
loguru==0.7.2

# 异步处理
asyncio
aiofiles==23.2.1

# 工具库
click==8.1.7
rich==13.6.0 